Notebook explores imagenet pretrained inception_v3
Basically used a pretrained inception_v3 and trained a 2 layer fully connected after for 20 epochs. The results on the test data were good, but grain of salt. I pulled the data out as a shuffle, so the model saw very similar images imagine it saw frame 1,2,3, and 5, for was the test image 4 would be very similar to 3 and 5.